requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
openai
google-auth
python-multipart
psycopg2-binary
sqlalchemy
eng_to_ipa
PyJWT
pytest
pytest-asyncio
httpx
tzdata